Believe that the random issues have been resolved. It finally dawned on me that I was removing readings from one thread and expecting that would always be available on the UI Thread. The solution was to move the remove and referencing readings on the same thread, the UI thread.
BTW: The errors where that items were outside of the bounds of the list. This was really confusing as a check right before verified that the index was in bounds. The reason that it was random is the removal on the connection thread had to remove the item right after the check but before it was referenced. This is why it may take several hours before the application crashed.